Add blob and short integration

+++ scc2425-tukano/src/main/java/tukano/impl/JavaBlobs.java
@@ -11,6 +11,8 @@ import utils.Hash;
 import utils.Hex;
 
 import static java.lang.String.format;
+import static tukano.api.Result.ErrorCode.FORBIDDEN;
+import static tukano.api.Result.error;
 
 public class JavaBlobs implements Blobs {
 
@@ -36,9 +38,8 @@ public class JavaBlobs implements Blobs {
     public Result<Void> upload(String blobId, byte[] bytes, String token) {
         Log.info(() -> format("upload : blobId = %s, sha256 = %s, token = %s\n", blobId, Hex.of(Hash.sha256(bytes)), token));
 
-//        ToDo add Validation
-//		if (!validBlobId(blobId, token))
-//			return error(FORBIDDEN);
+		if (!validBlobId(blobId, token))
+			return error(FORBIDDEN);
 
         return storage.write(toPath(blobId), bytes);
     }
@@ -47,9 +48,8 @@ public class JavaBlobs implements Blobs {
     public Result<byte[]> download(String blobId, String token) {
         Log.info(() -> format("download : blobId = %s, token=%s\n", blobId, token));
 
-//        ToDo add Validation
-//		if( ! validBlobId( blobId, token ) )
-//			return error(FORBIDDEN);
+		if( ! validBlobId( blobId, token ) )
+			return error(FORBIDDEN);
 
         return storage.read(toPath(blobId));
     }
@@ -58,9 +58,8 @@ public class JavaBlobs implements Blobs {
     public Result<Void> delete(String blobId, String token) {
         Log.info(() -> format("delete : blobId = %s, token=%s\n", blobId, token));
 
-//        ToDo add Validation
-//		if( ! validBlobId( blobId, token ) )
-//			return error(FORBIDDEN);
+		if( ! validBlobId( blobId, token ) )
+			return error(FORBIDDEN);
 
         return storage.delete(toPath(blobId));
     }
@@ -69,9 +68,8 @@ public class JavaBlobs implements Blobs {
     public Result<Void> deleteAllBlobs(String userId, String token) {
         Log.info(() -> format("deleteAllBlobs : userId = %s, token=%s\n", userId, token));
 
-//        ToDo add Validation
-//		if( ! Token.isValid( token, userId ) )
-//			return error(FORBIDDEN);
+		if( ! Token.isValid( token, userId ) )
+			return error(FORBIDDEN);
 
         return storage.delete(toPath(userId));
     }
